backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.0.300
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
torch>=2.0.0
